Update latest rust

+++ src/cardshoetracker.rs
@@ -4,7 +4,7 @@ use cards::card::Card;
 use shoe::shoe::DirectShoe;
 
 pub struct CardShoeTracker<'a> {
-    counts: [[uint;4];13],
+    counts: [[u32;4];13],
     wrapping: &'a mut (DirectShoe + 'a),
 }
 
@@ -17,10 +17,10 @@ impl <'a>DirectShoe for CardShoeTracker<'a> {
         };
         return r;
     }
-    fn len(&self) -> uint {
+    fn len(&self) -> usize {
         return self.wrapping.len();
     }
-    fn count(&self, v: &Value) -> uint {
+    fn count(&self, v: &Value) -> u32 {
         return self.wrapping.count(v);
     }
     fn remove(&mut self, v: &Value) -> Option<Card> {
@@ -29,10 +29,10 @@ impl <'a>DirectShoe for CardShoeTracker<'a> {
     fn insert(&mut self, v: &Card) {
         self.wrapping.insert(v);
     }
-    fn initial_length(&self) -> Option<uint> {
+    fn initial_length(&self) -> Option<u32> {
         self.wrapping.initial_length()
     }
-    fn maximum_count_of_any_value(&self) -> Option<uint> {
+    fn maximum_count_of_any_value(&self) -> Option<u32> {
         self.wrapping.maximum_count_of_any_value()
     }
 }
@@ -47,14 +47,14 @@ impl <'a>CardShoeTracker<'a> {
         };
     }
     */
-    pub fn count_value(&self, ref v: Value) -> uint {
+    pub fn count_value(&self, ref v: Value) -> u32 {
         let mut ret = 0;
         for &i in self.counts[v.index()].iter() {
             ret += i;
         }
         return ret;
     }
-    pub fn seen_cards(&self) -> uint {
+    pub fn seen_cards(&self) -> u32 {
         let mut ret = 0;
         for v in value::VALUES.iter() {
             ret += self.count_value(*v);
