+++ src/shoe/deck.rs
@@ -2,10 +2,10 @@ use cards::suit;
 use cards::card::Card;
 use cards::value;
 
-pub fn cards_in_deck<'a>(num_decks: uint) -> Vec<Card> {
+pub fn cards_in_deck<'a>(num_decks: u32) -> Vec<Card> {
     let mut ret = Vec::new();
     ret.clear();
-    ret.reserve_exact(52 * num_decks);
+    ret.reserve_exact(52 * num_decks as usize);
     for _ in range(0, num_decks) {
         for s in suit::SUITS.iter() {
             for v in value::VALUES.iter() {
